Implement a position-coin representation format
The building of the coin operator (in coin classes) and of the
interaction operator (in the DiscreteTimeQuantumWalk class)
now consider quantum systems in the position-coin representation format.
Also, this new representation format is now considered in the measurement
process of the system (in the state class).

The profiler class for quantum walks was renamed.